+++ src/main/webapp/src/Events.js
@@ -25,7 +25,9 @@
 
 let Events = Object.freeze({
   NEW_INFORMATION: 'new-information-message',
-  COLUMN_COUNT_RETRIEVED: 'column-count-retrieved'
+  COLUMN_COUNT_RETRIEVED: 'column-count-retrieved',
+  TOGGLED_HEALTH_DISPLAY: 'toggled-health-display',
+  TOGGLED_SUMMARY_DISPLAY: 'toggled-summary-display'
 });
 
 export {Events};