+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Change Log
 
 #### Next release
+- Modified run description logic
+- Made job health & run description display optional
 
 #### v0.5.0-build.1566142512 ~ 2019.08.18
 - Changed Jenkins logo on monitor